% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{bra.inaugural}
\alias{bra.inaugural}
\title{Brazilian Presidents' Inaugural Speeches}
\format{
a quanteda \link{corpus} object with the following variables:
\itemize{
\item \code{doc_id} character; Document ID
\item \code{text} character; Text of the inaugural speech
\item \code{President} character; Name of the President who gave the speech
\item \code{Legislature} numeric; The number of the legislature
\item \code{Date} date; Date of the Speech
\item \code{Military} logical; Military President
\item \code{Party} character; Political Party of the President
\item \code{Interrupted} logical; Was the administration interrupted?
\item \code{Interrupt.Cause} character; Cause of the interruption
\item \code{link_photo} character; URL for the photos of the President
}
}
\source{
Presidencia.gov.br and Bonfim (2005).
}
\usage{
bra.inaugural
}
\description{
This dataset contains 37 Inaugural Speeches from 1889 to 2023.
}
\examples{
\dontrun{
# some operations on the corpus
summary(bra.inaugural)
bra.inaugural
}
}
\keyword{data}
