% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{cis.corrupt}
\alias{cis.corrupt}
\title{Political Corruption in Spain}
\format{
a data.frame object with the following variables:
\itemize{
\item \code{Persona} character; The person speaking.
\item \code{text} character; Text containing the intervention of each participant.
\item \code{Order} integer; Order in the overall interventions.
\item \code{Tipo.Persona} character; Tipe of participant: Moderator, Female, or Male.
\item \code{Estudio} character; Study number: E2863.
\item \code{Grupo.Discusion} integer; Number of the focus group.
\item \code{Ciudad} character; Name of the city where the focus group took place.
\item \code{Fecha} date; Date of the focus group.
\item \code{Grupo.Demografico} character; Name of the demographic group: "Housewives", "Businessmen","University Students", "Public Servants", "Occupied - Young", "Workers", and "Liberal Professionals".
\item \code{Desc.Grupo} character; Detailed description of each demographic group.
\item \code{Edad} character; Age interval of participants.
}
}
\source{
CIS (2011).
}
\usage{
cis.corrupt
}
\description{
This dataset contains the transcription of 7 focus groups on political corruption in Spain. The study was performed by the Centro de Investigaciones Sociológicas (CIS) under the number E2863 in March, 2011.
}
\examples{
# some operations on the data
summary(cis.corrupt)
head(cis.corrupt)
}
\keyword{data}
