% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{dic.pol.es}
\alias{dic.pol.es}
\title{dic.pol.es - Spanish Political Discourses Dictionary}
\format{
a quanteda \link{dictionary} object with the following levels:
\itemize{
\item \code{level1} character; Broad political issues.
\item \code{level2} character; Specific codes.
\item \code{keywords} character; Terms to be employed in the search for each code/issue.
}
}
\source{
Own elaboration.
}
\usage{
dic.pol.es
}
\description{
This dictionary contains a nested (two-level) set of codes for analyzing Spanish inaugural speeches.
}
\examples{
# some operations on the corpus
summary(Pirandello)
dic.pol.es
}
\keyword{dictionary}
