\name{forceDirectedTree}
\alias{forceDirectedTree}
\title{Force Directed Tree Chart}
\description{
Creates a Force Directed Tree Chart for visualizing hierarchical data.
}
\usage{
forceDirectedTree(json_data, 
                  desc="units", 
                  attraction=-5, 
                  collapsed=FALSE, 
                  palette="Spectral", 
                  col.n=9, 
                  show.link=TRUE, 
                  url.return=FALSE, 
                  html.return=FALSE, 
                  viewer=TRUE, 
                  height=800)
}

\arguments{
\item{json_data}{A json string character as the one produced by the function jsonTree.}

\item{desc}{Description of the value units to be included in the tooltip.}

\item{attraction}{The intensity of attraction among nodes. The default is -5.}

\item{collapsed}{Logical. Indicates that all nodes should be collapsed. The default is FALSE.}

\item{palette}{A color palette contained in either RColorBrewer or wesanderson packages. The default is "Spectral".}

\item{col.n}{The number of colors to represent. The default is 9.}

\item{show.link}{Logical. Indicates that links connecting nodes will be visible. The default is TRUE.}

\item{url.return}{Logical. Indicates if the function should return the url address of the html file. The default is FALSE.}

\item{html.return}{Logical. Indicates if the function should return the html source code generated by the function. This option is useful for those wanting to save the results to a file. The default is FALSE.}

\item{viewer}{Logical. Indicates if the function should use RStudio viewer panel or the default browser to display the results. The default is TRUE.}

\item{height}{The height of the html panel in pixels. The default is 800.}
}
\details{
The function generates a Force Directed Tree Graph. It allows users to explore hierarchical data up to three levels.}
\value{
A Force Directed Tree chart or the html source code generated.
}
\examples{
\dontrun{
# Create a test dataset
dt <- data.frame(
          Characters=c("Joao Grilo","Chico"), 
          Word.Type= c("Verb","Noun"), 
          Word=c("sing","harmonica"), 
          Frequency=c(2,3))
          
# Call the function jsonTree to generate data
jso <- jsonTree(data=dt, 
         groups=c("Characters","Word.Type"),
         elements="Word",
         value="Frequency")

# Create the chart
forceDirectedTree(jso)
}
}
