% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{pal}
\alias{pal}
\title{Data Visualization Color Palettes}
\format{
An object of class \code{list} of length 222.
}
\usage{
pal
}
\description{
This list contains 222 color palettes classified by type of data. Each name repreents the type: cat - categorical, div - divergent, or seq - sequential; the source R package, the name of the palette and the number of hues or colors in the original palette.
}
\keyword{datasets}
