% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{Pirandello}
\alias{Pirandello}
\title{Pirandello - Six Characters in Search of an Author}
\format{
a \link{data.frame} object with the following variables:
\itemize{
\item \code{Order} numeric; The order of the speech in the play.
\item \code{Personaje} character; The character responsible for the speech.
\item \code{Action} character; Description of actions, emotions or gestures of the character immediately before his or her speech.
\item \code{text} character; The text of the speech.
}
}
\source{
Wikipedia: https://es.wikisource.org/wiki/Seis_personajes_en_busca_de_autor.
}
\usage{
Pirandello
}
\description{
This dataset contains the dialogues of the 1921 play Six Characters in Search of an Author by the Italian writer (and Nobel laureate) Luigi Pirandello. The Spanish version contained here was translated in 1926 by Félix Azzati and is currently available at Wikipedia.
}
\examples{
# some operations on the corpus
summary(Pirandello)
head(Pirandello)
}
\keyword{data}
