\name{plotChord}
\alias{plotChord}
\title{Co-ocurrence Chord Diagram for Dictionary Codes and Themes}
\description{
Generates Chord Diagram for the Co-Ocurrence of Dictionary Themes in a Corpus. 
}
\usage{
plotChord(data, 
           from="from", 
           to="to", 
           value="value", 
           font.size=12, 
           node.width=5, 
           opacity=0.05, 
           div.name="chordtheme", 
           height=600, 
           radius.percent=70, 
           viewer=TRUE, 
           html.return=FALSE, 
           url.return=FALSE)
}

\arguments{
\item{data}{A data.frame with three columns: origin, destination, and value.}

\item{from}{The name of the variable containing the information about the group, origin or aggregate measure. The default is "from".}

\item{to}{The name of the variable containing the information about the detail category, destination or attribute. The default is "to".}

\item{value}{The name of the variable containing the frequency or the value of the relationship between a given origin and destination or group and attribute. The default is "value".}

\item{font.size}{The size of the font in pixels. The default is 12.}

\item{node.width}{The width (thickness) of the node bar. The default is 5.}

\item{opacity}{The opacity of the links when not hovered. The default is 0.05 (5 percent opacity).}

\item{div.name}{Name of the div element employed to contain the graph. It is useful when you use the same type of graph multiple times in the same markdown page, for instance. The default is "chordtheme".}

\item{height}{The height of the html panel in pixels. The default is 580.}

\item{radius.percent}{Percentage of the height occupied by the radius of the diagram. The default is 70.}

\item{viewer}{Logical. Indicates if the function should use RStudio viewer panel or the default browser to display the results. The default is TRUE.}

\item{html.return}{Logical. Indicates if the function should return the html source code generated by the function. This option is useful for those wanting to save the results to a file. The default is FALSE.}

\item{url.return}{Logical. Indicates if the function should return the url address of the html file. The default is FALSE.}
}
\details{
The function generates an interactive Chord Diagram. It allows users to perform thematic analysis and other categorical hierarchical data analysis intuitively.}
\value{
A chart representing the similarity of texts or the html source code generated.
}
\examples{
\dontrun{
# Retrieve a corpus of text 
cp <- quanteda::corpus(spa.inaugural)

# Loads the dictionary for Spanish Speeches
dic <- dic.pol.es

# Generate the data to be employed
d <- matchCodes(cp, dic)

# Create the graph for the inaugural discourses
plotChord(d, from="term1", to="term2", value="value")

}
}
