% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{spa.inaugural}
\alias{spa.inaugural}
\title{Spanish Presidents' Inaugural Speeches}
\format{
a \link{data.frame} object with the following variables:
\itemize{
\item \code{doc_id} character; Document ID
\item \code{text} character; Text of the inaugural speech
\item \code{President} character; Name of the President who gave the speech
\item \code{Legislature} numeric; The number of the legislature
}
}
\source{
Moncloa.
}
\usage{
spa.inaugural
}
\description{
This dataset contains 15 Inaugural Speeches from 1979 to 2020.
}
\examples{
# some operations on the corpus
summary(spa.inaugural)
spa.inaugural
}
\keyword{data}
