% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{spa.sessions}
\alias{spa.sessions}
\title{Legislative Interventions on the 14th Legislature of the Spanish Congress (2019-2023)}
\format{
a data.frame object with the following variables:
\itemize{
\item \code{leg.number} integer; The number of the legislature.
\item \code{session.date} date; Date when the session was carried out.
\item \code{session.type} character; The type of session: "Plenaria" (floor) or "Diputación Permanente".
\item \code{session.number} integer; A sequential number assigned to identify the session.
\item \code{issue.type} character; The type of issues being discussed during proceedings (voting, debate, questioning and control).
\item \code{issue.details} character; Details about the issues or the questions addressed in the intervention of the representative.
\item \code{speech.order} integer; The number indicating the sequence of intervention in each session.
\item \code{speech.text} character; The complete text of the intervention.
\item \code{rep.name} character; The full name of the representative doing the intervention.
\item \code{rep.district} character; The electoral district of the representative.
\item \code{rep.party} character; The electoral party of the representative.
\item \code{rep.group} character; The parliamentary group of the representative.
\item \code{rep.condition} character; The "place of speech" of the representative (as a representative, a candidate, a state minister, as the president, as a member of the parliament board).
\item \code{rep.institution} character; The institution the representative belongs (the lower chamber, a particular ministry, the presidential office).
\item \code{speech.tokens} integer; The number of tokens of the speech.
}
}
\source{
Own elaboration based on the legislative diaries provided by the Congreso de los Diputados (www.congreso.es).
}
\usage{
spa.sessions
}
\description{
This dataset contains all interventions by lower-chamber representatives in the Spanish Congress (Congreso de los diputados) during the 14th legislature (2019-2023). Each observation corresponds to an intervention.
}
\examples{
# some operations on the corpus
summary(spa.sessions)
head(spa.sessions)
}
\keyword{Congress}
\keyword{Spanish}
\keyword{data.frame,}
