% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TENETCacheAllData.R
\name{TENETCacheAllData}
\alias{TENETCacheAllData}
\title{Cache all online datasets required by TENET examples and optional features}
\usage{
TENETCacheAllData()
}
\value{
Returns NULL.
}
\description{
This function locally caches all online TENET and SeSAMe datasets required
by TENET examples and optional features (TENET.ExperimentHub objects used in
examples, TENET.AnnotationHub datasets used in step 1, and SeSAMe datasets
loaded via the \code{DNAMethylationArray} argument). The main purpose of this
function is to enable the use of TENET in an environment without internet
access, such as the compute nodes of an HPC cluster. In this case, you must
run \code{TENETCacheAllData()} once while connected to the internet before using
TENET examples or these optional features.
}
\examples{
TENETCacheAllData()
}
