\name{themeChord}
\alias{themeChord}
\title{Co-ocurrence Chord Diagram for Dictionary Codes and Themes}
\description{
Generates Chord Diagram for the Co-Ocurrence of Dictionary Themes in a Corpus. 
}
\usage{
themeChord(corpus, 
           dic, 
           level=1, 
           remove.self=TRUE, 
           font.size=12, 
           node.width=5, 
           opacity=0.05, 
           div.name="chordtheme", 
           height=600, 
           radius.percent=70, 
           viewer=TRUE, 
           html.return=FALSE, 
           url.return=FALSE)
}

\arguments{
\item{corpus}{A quanteda corpus object.}

\item{dic}{Quanteda dictionary employed in the analysis.}

\item{level}{The level of the dictionary to take into account of the aggregation of co-ocurrences. The default is 1 (top level of abstraction).}

\item{remove.self}{Logical. Indicates if the function should remove co-ocurrences of a theme with itself. The default is TRUE.}

\item{font.size}{The size of the font in pixels. The default is 12.}

\item{node.width}{The width (thickness) of the node bar. The default is 5.}

\item{opacity}{The opacity of the links when not hovered. The default is 0.05 (5 percent opacity).}

\item{div.name}{Name of the div element employed to contain the graph. It is useful when you use the same type of graph multiple times in the same markdown page, for instance. The default is "chordtheme".}

\item{height}{The height of the html panel in pixels. The default is 580.}

\item{radius.percent}{Percentage of the height occupied by the radius of the diagram. The default is 70.}

\item{viewer}{Logical. Indicates if the function should use RStudio viewer panel or the default browser to display the results. The default is TRUE.}

\item{html.return}{Logical. Indicates if the function should return the html source code generated by the function. This option is useful for those wanting to save the results to a file. The default is FALSE.}

\item{url.return}{Logical. Indicates if the function should return the url address of the html file. The default is FALSE.}
}
\details{
The function generates a Force Directed Hierarchical Cluster Tree Network Graph for quanteda corpuses. It calculates the similarity among texts and organizes them according to a tree structure. The interactive traits of the chart allow users to explore these relationships with more depth.}
\value{
A chart representing the similarity of texts or the html source code generated.
}
\examples{
\dontrun{
# Retrieve a corpus of text 
cp <- quanteda::corpus(spa.inaugural)

# Loads the dictionary for Spanish Speeches
dic <- dic.pol.es

# Create the graph for the inaugural discourses
themeChord(cp,dic)

# Reshapes to sentences and use only the second level
# of the dictionary
cs <- corpus_reshape(cp, "sentences")

themeChord(cp,dic, level=2)
}
}
