\name{forceClusTree}
\alias{forceClusTree}
\title{Force Directed Hierarchical Cluster Tree Network}
\description{
Generates a Force Directed Hierarchical Cluster Tree Network Graph for corpus objects. 
}
\usage{
forceClusTree(corpus, 
              remove_punct=TRUE,
              lang = "es",
              url.return = FALSE,
              html.return = FALSE,
              viewer = TRUE,
              custom.color = NULL,
              palette = "BottleRocket2",
              custom.pal = NULL,
              groupvar = NULL,
              height=580,
              link.col="#734421",
              link.width=2,
              minRadius=0,
              maxRadius=100,
              BodyStrength=-10,
              include.text=FALSE,
              img.docvar=NULL,
              clust.method="euclidean",
              div.name="chartdivclus")
}

\arguments{
\item{corpus}{A quanteda corpus object.}

\item{remove_punct}{Remove punctuation. The default is TRUE.}

\item{lang}{The language for removing stopwords. The default is Spanish: "es".}

\item{url.return}{Logical. Indicates if the function should return the url address of the html file. The default is FALSE.}

\item{html.return}{Logical. Indicates if the function should return the html source code generated by the function. This option is useful for those wanting to save the results to a file. The default is FALSE.}

\item{viewer}{Logical. Indicates if the function should use RStudio viewer panel or the default browser to display the results. The default is TRUE.}

\item{custom.color}{The color used for coloring nodes if all should be colored with only one color. The default is NULL.}

\item{palette}{The color palette included in either RColorBrewer or wesanderson palettes. The default is "BottleRocket2".}

\item{custom.pal}{The list of colors used for creating a customized color palette. The default is NULL.}

\item{groupvar}{Metadata variable associated to the texts on a quanteda corpus object indicating groups or more aggregated categories.}

\item{height}{The height of the html panel in pixels. The default is 580.}

\item{link.col}{The color of the links connecting nodes in the network. The default is "#734421" (dark brown).}

\item{link.width}{The width of the links connecting nodes in the network. The default is 2.}

\item{minRadius}{The minimum radius of the nodes in the network. The default is 0.}

\item{maxRadius}{The maximum radius of the nodes in the network. The default is 100.}

\item{BodyStrength}{The attraction force of nodes in the network.  The default is -10.}

\item{include.text}{Logical. Indicates if the function should include the text in the tooltip. It should be used only with small texts. Large texts would collapse the visualization. The default is FALSE.}

\item{img.docvar}{Indicates the name of the variable in the corpus metadata containing the url links to the images to be displayed in the tooltip. The default is NULL.}

\item{clust.method}{Indicates the method for calculating the text distance. The options are: "euclidean" (default), "manhattan", "maximum", "canberra", and "minkowski".}

\item{div.name}{Name of the div element employed to contain the graph. It is useful when you use the same type of graph multiple times in the same markdown page, for instance. The default is "chartdiv".}

}
\details{
The function generates a Force Directed Hierarchical Cluster Tree Network Graph for quanteda corpuses. It calculates the similarity among texts and organizes them according to a tree structure. The interactive traits of the chart allow users to explore these relationships with more depth.}
\value{
A chart representing the similarity of texts or the html source code generated.
}
\examples{
\dontrun{
# Retrieve a corpus of text 
tx <- quanteda::data_corpus_inaugural

# Create the graph for the inaugural discourses
forceClusTree(tx, lang="en", maxRadius = 50)
}
}
