% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step7LinkedDNAMethylationSiteCountHistograms.R
\name{step7LinkedDNAMethylationSiteCountHistograms}
\alias{step7LinkedDNAMethylationSiteCountHistograms}
\title{Create histograms displaying the number of genes or transcription factors
linked to a given number of RE DNA methylation sites}
\usage{
step7LinkedDNAMethylationSiteCountHistograms(
  TENETMultiAssayExperiment,
  hypermethGplusAnalysis = TRUE,
  hypomethGplusAnalysis = TRUE
)
}
\arguments{
\item{TENETMultiAssayExperiment}{Specify a MultiAssayExperiment object
containing expression and methylation SummarizedExperiment objects, such as
one created by the TCGADownloader function. This MultiAssayExperiment object
should also contain the results from the
step6DNAMethylationSitesPerGeneTabulation function in its metadata.}

\item{hypermethGplusAnalysis}{Set to TRUE to create histograms of genes/TFs
linked to hypermethylated RE DNA methylation sites with G+ links. Defaults
to TRUE.}

\item{hypomethGplusAnalysis}{Set to TRUE to create histograms of genes/TFs
linked to hypomethylated RE DNA methylation sites with G+ links. Defaults to
TRUE.}
}
\value{
Returns the MultiAssayExperiment object given as the
TENETMultiAssayExperiment argument with an additional list of information
named 'step7LinkedDNAMethylationSiteCountHistograms' in its metadata with
the output of this function. This list is subdivided into hypermethGplus or
hypomethGplus results as selected by the user. Each of these contains a
histogram displaying the number of all genes, or TFs only, linked to a given
number of RE DNA methylation sites in each of the selected analysis
quadrants.
}
\description{
This function generates histograms displaying the number of all genes,
or transcription factor (TF) genes only, with links to a given number of
regulatory element DNA methylation sites from both of the hyper- or
hypomethylated G+ analysis quadrants, as selected by the user.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
## This example uses the example MultiAssayExperiment provided in the
## TENET.ExperimentHub package to create histograms displaying counts of
## genes/TFs by the number of linked hyper- or hypomethylated G+ RE DNA
## methylation sites.

## Load the example TENET MultiAssayExperiment object
## from the TENET.ExperimentHub package
exampleTENETMultiAssayExperiment <-
    TENET.ExperimentHub::exampleTENETMultiAssayExperiment()

## Use the example dataset to create the RE DNA methylation site count
## histograms
returnValue <- step7LinkedDNAMethylationSiteCountHistograms(
    TENETMultiAssayExperiment = exampleTENETMultiAssayExperiment
)

## This example also uses the example MultiAssayExperiment provided in the
## TENET.ExperimentHub package, but only creates a histogram of counts of
## genes/TFs by the number of linked hypomethylated G+ RE DNA methylation
## sites.

## Load the example TENET MultiAssayExperiment object
## from the TENET.ExperimentHub package
exampleTENETMultiAssayExperiment <-
    TENET.ExperimentHub::exampleTENETMultiAssayExperiment()

## Use the example dataset to create the RE DNA methylation site count
## histograms
returnValue <- step7LinkedDNAMethylationSiteCountHistograms(
    TENETMultiAssayExperiment = exampleTENETMultiAssayExperiment,
    hypermethGplusAnalysis = FALSE
)
\dontshow{\}) # examplesIf}
}
