% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step7LinkedDNAMethylationSitesMotifSearching.R
\name{step7LinkedDNAMethylationSitesMotifSearching}
\alias{step7LinkedDNAMethylationSitesMotifSearching}
\title{Perform motif searching for transcription factor motifs in the vicinity of
RE DNA methylation sites linked to the specified transcription factors}
\usage{
step7LinkedDNAMethylationSitesMotifSearching(
  TENETMultiAssayExperiment,
  TFMotifList,
  geneAnnotationDataset = NA,
  DNAMethylationArray = NA,
  hypermethGplusAnalysis = TRUE,
  hypomethGplusAnalysis = TRUE,
  distanceFromREDNAMethylationSites = 100,
  matchPWMMinScore = "75\%",
  coreCount = 1
)
}
\arguments{
\item{TENETMultiAssayExperiment}{Specify a MultiAssayExperiment object
containing expression and methylation SummarizedExperiment objects, such as
one created by the TCGADownloader function. This MultiAssayExperiment object
should also contain the results from the \code{step5OptimizeLinks} function
in its metadata.}

\item{TFMotifList}{Specify a named list mapping transcription factor gene
names and/or IDs to their respective motif position weight matrix (PWM). The
PWMs should be in the form of a 4xN matrix. If a PWM is not available, the
user can select one by querying the MotifDb package database. See the
example below for more infornation. \strong{Note:} The name of each PWM in the
list must match the gene name or Ensembl ID of a gene in the
TENETMultiAssayExperiment with RE DNA methylation sites linked to it for the
specified analysis types.}

\item{geneAnnotationDataset}{Specify a gene annotation dataset which is
used to identify names for genes by their Ensembl IDs. The argument must be
either a GRanges object (such as one imported via \code{rtracklayer::import}) or a
path to a GFF3 or GTF file. Both GENCODE and Ensembl annotations are
supported. Other annotation datasets may work, but have not been tested.
See the "Input data" section of the vignette for information on the required
dataset format.
Specify NA to use the names for genes listed in the "geneName" column of the
elementMetadata of the rowRanges of the "expression" SummarizedExperiment
object within the TENETMultiAssayExperiment object. Defaults to NA.}

\item{DNAMethylationArray}{Specify the name of a DNA methylation probe array
supported by the sesameData package (see
\code{?sesameData::sesameData_getManifestGRanges}). If an array is specified, RE
DNA methylation sites and their locations in that array's manifest are
cross-referenced with RE DNA methylation site IDs included in the rownames
of the methylation dataset provided in the "methylation"
SummarizedExperiment object within the TENETMultiAssayExperiment object, and
only those overlapping will be considered for analysis. If set to NA, all RE
DNA methylation sites with locations listed in the rowRanges of the
"methylation" SummarizedExperiment object are used. Defaults to NA.}

\item{hypermethGplusAnalysis}{Set to TRUE to do motif searching in the
vicinity of hypermethylated RE DNA methylation sites with G+ links to the TF
of interest. Defaults to TRUE.}

\item{hypomethGplusAnalysis}{Set to TRUE to do motif searching in the
vicinity of hypomethylated RE DNA methylation sites with G+ links to the TF
of interest. Defaults to TRUE.}

\item{distanceFromREDNAMethylationSites}{Specify a positive integer in base
pairs to be the distance from the linked RE DNA methylation sites within
which motif searching will be performed. Defaults to 100.}

\item{matchPWMMinScore}{Specify the \code{min.score} argument passed to the
matchPWM function for motif searching. See \code{?Biostrings::matchPWM} for more
details. Defaults to "75\%".}

\item{coreCount}{Argument passed as the mc.cores argument to mclapply. See
\code{?parallel::mclapply} for more details. Defaults to 1.}
}
\value{
Returns the MultiAssayExperiment object given as the
TENETMultiAssayExperiment argument with an additional list of information
named 'step7LinkedDNAMethylationSitesMotifSearching' in its metadata with
the output of this function. This list is subdivided into hypermethGplus or
hypomethGplus results as selected by the user. Each contains a list for each
specified TF gene with a seqLogo of the gene's motif PWM, a table of the
PWM, and two additional tables with information on motif occurrences found
in the vicinity of linked RE DNA methylation sites. The first table contains
information about each motif found in the vicinity of the linked RE DNA
methylation sites for the TF gene at the specified matchPWMMinScore. The
second lists each RE DNA methylation site linked to the TF gene for the
analysis types specified, and the number of motif occurrences found in the
vicinity of each of those RE DNA methylation sites.
}
\description{
This function takes a named list of transcription factors (TFs) and their
binding motifs, and for each of the TFs, it identifies if the specified
motif for that TF is found within a user-specified distance to each RE DNA
methylation site linked to that TF from both of the hyper- or hypomethylated
G+ analysis quadrants, as selected by the user.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
## These examples require the MotifDb package
if (requireNamespace("MotifDb", quietly = TRUE)) {
    ## Show available motifs for example TF FOXA1
    names(MotifDb::query(MotifDb::MotifDb, "FOXA1"))

    ## The seqLogos for all input motifs will also be included in the output
    ## of this function. Alternatively, individual motifs can be visualized
    ## with the seqLogo function from the seqLogo package.
    seqLogo::seqLogo(MotifDb::query(MotifDb::MotifDb, "FOXA1")[[3]])

    ## Once PWMs have been selected for use, a list containing them must be
    ## created
    exampleTFMotifList <- list(
        "FOXA1" = MotifDb::query(MotifDb::MotifDb, "FOXA1")[[3]],
        "ESR1" = MotifDb::query(MotifDb::MotifDb, "ESR1")[[4]]
    )

    ## This example uses the example MultiAssayExperiment provided in the
    ## TENET.ExperimentHub package to perform motif overlapping for all
    ## hyper- and hypomethylated RE DNA methylation sites linked to the
    ## FOXA1 and ESR1 genes using the motifs for each gene specified in the
    ## exampleTFMotifList. Gene names and locations, and the locations of RE
    ## DNA methylation sites, will be retrieved from the rowRanges of the
    ## 'expression' and 'methylation' SummarizedExperiment objects in the
    ## example MultiAssayExperiment. Regions within 100 bp of linked RE DNA
    ## methylation sites will be checked for motifs, and a similarity
    ## threshold of 75\% will be used to identify motifs. The analysis will
    ## be performed using one CPU core.

    ## Load the example TENET MultiAssayExperiment object
    ## from the TENET.ExperimentHub package
    exampleTENETMultiAssayExperiment <-
        TENET.ExperimentHub::exampleTENETMultiAssayExperiment()

    ## Use the example dataset to do the motif searching
    returnValue <- step7LinkedDNAMethylationSitesMotifSearching(
        TENETMultiAssayExperiment = exampleTENETMultiAssayExperiment,
        TFMotifList = exampleTFMotifList
    )

    ## This example uses the example MultiAssayExperiment provided in the
    ## TENET.ExperimentHub package to perform motif overlapping for only
    ## hypomethylated RE DNA methylation sites linked to the FOXA1 and ESR1
    ## genes using the motifs for each gene specified in the
    ## exampleTFMotifList. Gene names and locations, and the locations of RE
    ## DNA methylation sites, will be retrieved from the rowRanges of the
    ## 'expression' and 'methylation' SummarizedExperiment objects in the
    ## example MultiAssayExperiment. Regions within 50 bp of linked RE DNA
    ## methylation sites will be checked for motifs, and a similarity
    ## threshold of 80\% will be used to identify motifs. The analysis will
    ## be performed using 8 CPU cores.

    ## Load the example TENET MultiAssayExperiment object
    ## from the TENET.ExperimentHub package
    exampleTENETMultiAssayExperiment <-
        TENET.ExperimentHub::exampleTENETMultiAssayExperiment()

    ## Use the example dataset to do the motif searching
    returnValue <- step7LinkedDNAMethylationSitesMotifSearching(
        TENETMultiAssayExperiment = exampleTENETMultiAssayExperiment,
        TFMotifList = exampleTFMotifList,
        hypermethGplusAnalysis = FALSE,
        distanceFromREDNAMethylationSites = 50,
        matchPWMMinScore = "80\%",
        coreCount = 8
    )
} else {
    message(
        "These examples require the MotifDb package. Run ",
        'BiocManager::install("MotifDb") and rerun these examples.'
    )
}
\dontshow{\}) # examplesIf}
}
