% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/step7SelectedDNAMethylationSitesCaseVsControlBoxplots.R
\name{step7SelectedDNAMethylationSitesCaseVsControlBoxplots}
\alias{step7SelectedDNAMethylationSitesCaseVsControlBoxplots}
\title{Generate boxplots comparing the methylation level of the specified RE DNA
methylation sites in case and control samples}
\usage{
step7SelectedDNAMethylationSitesCaseVsControlBoxplots(
  TENETMultiAssayExperiment,
  DNAMethylationSites,
  coreCount = 1
)
}
\arguments{
\item{TENETMultiAssayExperiment}{Specify a MultiAssayExperiment object
containing a methylation SummarizedExperiment object, such as one created by
the TCGADownloader function.}

\item{DNAMethylationSites}{Supply a vector of RE DNA methylation site IDs
for which boxplots with the methylation of those RE DNA methylation sites
will be created.}

\item{coreCount}{Argument passed as the mc.cores argument to mcmapply. See
\code{?parallel::mcmapply} for more details. Defaults to 1.}
}
\value{
Returns the MultiAssayExperiment object given as the
TENETMultiAssayExperiment argument with an additional list of data named
'step7SelectedDNAMethylationSitesCaseVsControlBoxplots' in its metadata with
the output of this function, which contains boxplots showing the methylation
of the RE DNA methylation sites of interest in the case and control samples,
with Student's t-test p-values and the ID of the RE DNA methylation site in
the title.
}
\description{
This function takes a vector of RE DNA methylation sites specified by the
user and generates boxplots displaying the methylation level of each of these
DNA methylation sites in the case compared to control samples, along with the
results of a Student's t-test comparing the methylation level between these
two groups.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
## This example uses the example MultiAssayExperiment provided in the
## TENET.ExperimentHub package to generate boxplots for several selected
## RE DNA methylation sites, using one CPU core.

## Load the example TENET MultiAssayExperiment object
## from the TENET.ExperimentHub package
exampleTENETMultiAssayExperiment <-
    TENET.ExperimentHub::exampleTENETMultiAssayExperiment()

## Use the example dataset to create RE DNA methylation site case vs. control
## boxplots
returnValue <- step7SelectedDNAMethylationSitesCaseVsControlBoxplots(
    TENETMultiAssayExperiment = exampleTENETMultiAssayExperiment,
    DNAMethylationSites = c("cg03095778", "cg24011501", "cg12989041"),
    coreCount = 1
)
\dontshow{\}) # examplesIf}
}
