% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step7StatesForLinks.R
\name{step7StatesForLinks}
\alias{step7StatesForLinks}
\title{Identify which of the case samples harbor each of the identified regulatory
element DNA methylation site-gene links}
\usage{
step7StatesForLinks(
  TENETMultiAssayExperiment,
  hypermethGplusAnalysis = TRUE,
  hypomethGplusAnalysis = TRUE,
  coreCount = 1
)
}
\arguments{
\item{TENETMultiAssayExperiment}{Specify a MultiAssayExperiment object
containing expression and methylation SummarizedExperiment objects, such as
one created by the TCGADownloader function. This MultiAssayExperiment object
should also contain the results from the \code{step5OptimizeLinks} functions in
its metadata.}

\item{hypermethGplusAnalysis}{Set to TRUE to analyze which case samples
harbor each hypermethylated G+ RE DNA methylation site-gene link. Defaults to
TRUE.}

\item{hypomethGplusAnalysis}{Set to TRUE to analyze which case samples
harbor each hypomethylated G+ RE DNA methylation site-gene link. Defaults to
TRUE.}

\item{coreCount}{Argument passed as the mc.cores argument to mcmapply. See
\code{?parallel::mcmapply} for more details. Defaults to 1.}
}
\value{
Returns the MultiAssayExperiment object given as the
TENETMultiAssayExperiment argument with an additional list of information
named 'step7StatesForLinks' in its metadata with the output of this
function. This list is subdivided into hypermethGplus or hypomethGplus
results as selected by the user, which contain data frames for each of the
specified analysis types with case samples in the columns and each RE DNA
methylation site-gene link for that analysis type in the rows, with a 1
indicating the sample is positive for that link and a 0 indicating it is not.
NA values are shown for samples that lack methylation data for the site or
expression data for the gene.
}
\description{
This function generates data frames with information for each of the case
samples and each RE DNA methylation site-gene link from both of the
hyper- or hypomethylated G+ analysis quadrants, as selected by the user, on
if a given sample is said to "harbor" each link, depending on if the
methylation of the given sample for the RE DNA methylation site in the link
is above or below the hyper- or hypomethylation cutoff defined in the
\code{step2GetDifferentiallyMethylatedSites} function and the expression of the
gene in the link is significantly greater than, or less than, the mean
expression in the control samples, as determined by a Bonferroni-corrected
1-sided t-test with a p-value threshold of 0.05.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
## This example uses the example MultiAssayExperiment provided in the
## TENET.ExperimentHub package to create data frames for link states between
## genes and both hyper- and hypomethylated G+ RE DNA methylation sites. The
## analysis will be performed using one CPU core.

## Load the example TENET MultiAssayExperiment object
## from the TENET.ExperimentHub package
exampleTENETMultiAssayExperiment <-
    TENET.ExperimentHub::exampleTENETMultiAssayExperiment()

## Use the example dataset to do the states for links analysis
returnValue <- step7StatesForLinks(
    TENETMultiAssayExperiment = exampleTENETMultiAssayExperiment
)

## This example uses the example MultiAssayExperiment provided in the
## TENET.ExperimentHub package to create data frames for link states between
## genes and only hypomethylated G+ RE DNA methylation sites. The analysis
## will be performed using 8 CPU cores.

## Load the example TENET MultiAssayExperiment object
## from the TENET.ExperimentHub package
exampleTENETMultiAssayExperiment <-
    TENET.ExperimentHub::exampleTENETMultiAssayExperiment()

## Use the example dataset to do the states for links analysis
returnValue <- step7StatesForLinks(
    TENETMultiAssayExperiment = exampleTENETMultiAssayExperiment,
    hypermethGplusAnalysis = FALSE,
    coreCount = 8
)
\dontshow{\}) # examplesIf}
}
