% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step7TopGenesTADTables.R
\name{step7TopGenesTADTables}
\alias{step7TopGenesTADTables}
\title{Create tables using user-supplied topologically associating domain (TAD)
information which identify the topologically associating domains
containing each RE DNA methylation site linked to the top genes/transcription
factors, as well as other genes in the same topologically associating domain
as potential downstream targets}
\usage{
step7TopGenesTADTables(
  TENETMultiAssayExperiment,
  TADFiles,
  geneAnnotationDataset = NA,
  DNAMethylationArray = NA,
  hypermethGplusAnalysis = TRUE,
  hypomethGplusAnalysis = TRUE,
  topGeneNumber = 10,
  coreCount = 1
)
}
\arguments{
\item{TENETMultiAssayExperiment}{Specify a MultiAssayExperiment object
containing expression and methylation SummarizedExperiment objects, such as
one created by the TCGADownloader function. This MultiAssayExperiment object
should also contain the results from the \code{step5OptimizeLinks} and
\code{step6DNAMethylationSitesPerGeneTabulation} functions in its metadata.}

\item{TADFiles}{Specify a data frame, matrix, or GRanges object with
information on the TAD compartments of interest, organized in a bed-like
manner (see \url{https://genome.ucsc.edu/FAQ/FAQformat.html#format1}), or a path
to a directory that contains bed-like files that contain such TAD
information. If a path is provided, multiple bed-formatted TAD files can be
included in the specified directory. The files may optionally be compressed
(.gz/.bz2/.xz). \strong{Note:} Data frames and matrices will be assumed to use
1-indexed coordinates; \code{rtracklayer::import.bed} converts coordinates to
1-indexed upon import.}

\item{geneAnnotationDataset}{Specify a gene annotation dataset which is
used to identify names for genes by their Ensembl IDs. The argument must be
either a GRanges object (such as one imported via \code{rtracklayer::import}) or a
path to a GFF3 or GTF file. Both GENCODE and Ensembl annotations are
supported. Other annotation datasets may work, but have not been tested.
See the "Input data" section of the vignette for information on the required
dataset format.
Specify NA to use the names for genes listed in the "geneName" column of the
elementMetadata of the rowRanges of the "expression" SummarizedExperiment
object within the TENETMultiAssayExperiment object. Defaults to NA.}

\item{DNAMethylationArray}{Specify the name of a DNA methylation probe array
supported by the sesameData package (see
\code{?sesameData::sesameData_getManifestGRanges}). If an array is specified, RE
DNA methylation sites and their locations in that array's manifest are
cross-referenced with RE DNA methylation site IDs included in the rownames
of the methylation dataset provided in the "methylation"
SummarizedExperiment object within the TENETMultiAssayExperiment object, and
only those overlapping will be considered for analysis. If set to NA, all RE
DNA methylation sites with locations listed in the rowRanges of the
"methylation" SummarizedExperiment object are used. Defaults to NA.}

\item{hypermethGplusAnalysis}{Set to TRUE to create TAD tables for the RE
DNA methylation sites linked to the top genes/TFs by most hypermethylated RE
DNA methylation sites with G+ links.}

\item{hypomethGplusAnalysis}{Set to TRUE to create TAD tables for the RE DNA
methylation sites linked to the top genes/TFs by most hypomethylated RE DNA
methylation sites with G+ links.}

\item{topGeneNumber}{Specify the number of top genes/TFs, based on the most
linked RE DNA methylation sites of a given analysis type, for which to
generate TAD tables for the RE DNA methylation sites linked to those genes.
Defaults to 10.}

\item{coreCount}{Argument passed as the mc.cores argument to mcmapply. See
\code{?parallel::mcmapply} for more details. Defaults to 1.}
}
\value{
Returns the MultiAssayExperiment object given as the
TENETMultiAssayExperiment argument with an additional list of information
named 'step7TopGenesTADTables' in its metadata with the output of this
function. This list is subdivided into hypermethGplus or hypomethGplus
results as selected by the user, which are further subdivided into data
frames with data for the unique RE DNA methylation sites linked to the top
overall genes, and for top TF genes only. This includes the top genes/TFs
each RE DNA methylation site is linked to, and, for each TAD file, if an RE
DNA methylation site was found in a TAD in that file, as well as the gene
count and identities of other genes found in the same TAD as each RE DNA
methylation site.
}
\description{
This function takes the top genes/transcription factors (TFs) by number of
linked RE DNA methylation sites identified by the
\code{step6DNAMethylationSitesPerGeneTabulation} function up to the number
specified by the user and generates tables with information for each of the
RE DNA methylation sites linked to them for both of the hyper- or
hypomethylated G+ analysis quadrants, as selected by the user. These tables
note which of the top genes/TFs each RE DNA methylation site is linked to,
as well as the total number and names of genes which happen to lie within
the same topologically associating domain (TAD) of each RE DNA methylation
site in each of the user-supplied TAD files.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
## This example uses the example MultiAssayExperiment provided in the
## TENET.ExperimentHub package to do overlapping for all unique RE DNA
## methylation sites linked to the top 10 genes, by number of linked hyper-
## or hypomethylated RE DNA methylation sites, using a GRanges object
## containing topologically associating domain (TAD) data from the
## TENET.ExperimentHub package. Gene names and locations, and the locations
## of RE DNA methylation sites, will be retrieved from the rowRanges of the
## 'expression' and 'methylation' SummarizedExperiment objects in the
## example MultiAssayExperiment. The analysis will be performed using one
## CPU core.

## Load the example TENET MultiAssayExperiment object
## from the TENET.ExperimentHub package
exampleTENETMultiAssayExperiment <-
    TENET.ExperimentHub::exampleTENETMultiAssayExperiment()

## Load the example TAD GRanges object from the TENET.ExperimentHub package
exampleTENETTADRegions <- TENET.ExperimentHub::exampleTENETTADRegions()

## Use the example datasets to do the TAD overlapping
returnValue <- step7TopGenesTADTables(
    TENETMultiAssayExperiment = exampleTENETMultiAssayExperiment,
    TADFiles = exampleTENETTADRegions
)

## This example uses the example MultiAssayExperiment provided in the
## TENET.ExperimentHub package to do overlapping for all unique RE DNA
## methylation sites linked to the top 5 genes only, by number of linked
## hypomethylated RE DNA methylation sites only, with bed-like files
## containing topologically associating domain (TAD) data located in the
## user's R working directory. This analysis will be done using gene names
## and their locations which are included in the "geneName" column of the
## elementMetadata of the rowRanges of the "expression" SummarizedExperiment
## object within the TENETMultiAssayExperiment object, and RE DNA
## methylation sites and their locations which are included in the HM450
## array retrieved via the sesameData package. The analysis will be
## performed using 8 CPU cores.

## Load the example TENET MultiAssayExperiment object
## from the TENET.ExperimentHub package
exampleTENETMultiAssayExperiment <-
    TENET.ExperimentHub::exampleTENETMultiAssayExperiment()

## Use the example dataset to do the TAD overlapping
returnValue <- step7TopGenesTADTables(
    TENETMultiAssayExperiment = exampleTENETMultiAssayExperiment,
    TADFiles = ".",
    DNAMethylationArray = "HM450",
    hypermethGplusAnalysis = FALSE,
    topGeneNumber = 5,
    coreCount = 8
)
\dontshow{\}) # examplesIf}
}
