\name{tagText}
\alias{tagText}
\title{Text Tagger}
\description{
The function tags text according to keywords or quanteda dictionaries.
}
\usage{

tagText(text, 
        title="Document Title", 
        keywords=NULL,
        dic=NULL,
        col="red", 
        bright=130, 
        tooltip=TRUE, 
        html.return=FALSE, 
        viewer=TRUE, 
        url.return=FALSE, 
        margin=50, 
        font.size=12)
}

\arguments{
\item{text}{Text to be used for tagging or highlighting.}
\item{title}{The title or header of the tagged document. The default is "Document Title".}
\item{keywords}{The keywords to be used to highlight or tag the submitted text. The default is NULL.}
\item{dic}{The name of the dictionary to be employed to highlight or tag the text. The default is NULL.}
\item{col}{The color of the text to be highlighted. It must the either hexadecimal or a javascript default color name. The default is "red".}
\item{bright}{Luminosity threshold to separate bright to dark colors. This parameter will be passed to the colBright function. The default is 130.}
\item{tooltip}{Logical. Indicates whether the final document should include a tooltip containing the dictionary categories or not. The default is TRUE.}
\item{html.return}{Logical. When TRUE, returns the html code generated by the function to be stored and saved into a file. The default is FALSE.}
\item{viewer}{Logical. When TRUE, show the results on RStudio Viewer panel. When FALSE, opens the file on the default browser. The default is TRUE.}
\item{url.return}{Logical. When TRUE, returns the location of the temporary html file stored in the computer. The default is FALSE.}
\item{margin}{Defines the size of the html file margins in pixels. The default is 50.}
\item{font.size}{Establishes the font size for the text to be represented. The default is 12.}
}
\details{
The function tagText finds and highlights words and expressions in a given text. It was designed to help qualitative researchers to locate themes and issues in narrative texts and documents. Therefore, it is meant to be an exploratory tool for qualitative analysis and a step in a broader coding process of text.
}
\value{
There are two possible values to be returned. 

The first is the html code generated by the function. This option is particularly interesting in the case users plan to save the results into a file. It becomes effective only when html.return=TRUE.

The second is the location of the temporary file generated by the function. It becomes effective when html.return=TRUE.


}
\examples{
\dontrun{
# Create a text variable as example
tx <- c("Political regimes can be classified 
as liberal democracies, electoral democracies, 
electoral autocracies, and full autocracies.")

# Find words starting with "democ" in the text
tagText(tx, keywords="democ")

library(quanteda)

# Creates a dictionary to classify regimes
reg <- dictionary(list(
                  Democracy=c("democ"),
                  Autocracy=c("autoc")
                      ))

# Change the default colors
tagText(tx, dic=reg, col=c("red","blue"))
}
}
