% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCGADownloader.R
\name{TCGADownloader}
\alias{TCGADownloader}
\title{Download TCGA gene expression, DNA methylation, and clinical datasets
and compile them into a MultiAssayExperiment object}
\usage{
TCGADownloader(
  rawDataDownloadDirectory,
  GDCDownloadMethod = "api",
  filesPerChunk = 10,
  TCGAStudyAbbreviation,
  RNASeqWorkflow,
  RNASeqLog2Normalization = TRUE,
  removeDupTumor = TRUE,
  matchingExpAndMetSamples = TRUE,
  clinicalSurvivalData = "combined",
  outputFile = NA
)
}
\arguments{
\item{rawDataDownloadDirectory}{Specify the path to the directory where
TCGAbiolinks should download data. Note that this dataset can be very
sizable.}

\item{GDCDownloadMethod}{The method to use when downloading data from the
Genomic Data Commons (GDC). Passed as the \code{method} argument to TCGAbiolinks'
\code{GDCdownload} function. The available options are "api" and "client"; the
default is "api". The "api" method works on all operating systems, but
it does not retry the download of incomplete or corrupted files, so
\code{TCGADownloader} must be manually rerun in this case. The "client" method is
more reliable, but it requires Windows, macOS (Apple Silicon only), or Ubuntu
(64-bit x86 only), or manual installation of the GDC Data Transfer Tool
Client (which must be in the command search path).}

\item{filesPerChunk}{The number of data files to download at once when using
the "api" download method. Passed as the \code{files.per.chunk} argument to
TCGAbiolinks' \code{GDCdownload} function. Lower values may improve download
reliability, but higher values may increase download speed. Defaults to 10.}

\item{TCGAStudyAbbreviation}{Input a four-letter code for a TCGA dataset
for which to download data. See
\url{https://gdc.cancer.gov/resources-tcga-users/tcga-code-tables/tcga-study-abbreviations}
for more information and a complete list of options.}

\item{RNASeqWorkflow}{Select the type of RNA-seq data to download. For
TENET purposes, choose either "STAR - FPKM", "STAR - FPKM-UQ",
"STAR - FPKM-UQ - old formula", or "STAR - TPM". "STAR - Counts" can also
be used but is not recommended for TENET analyses. See
\url{https://docs.gdc.cancer.gov/Data/Bioinformatics_Pipelines/Expression_mRNA_Pipeline/}
for more information on these.}

\item{RNASeqLog2Normalization}{Set to TRUE to do log2 normalization of
RNA-seq expression values. Defaults to TRUE.}

\item{removeDupTumor}{Set to TRUE to remove duplicate tumor samples
taken from the same subject, leaving only one sample per subject in
alphanumeric order. \strong{Note:} To properly create a dataset for use with
TENET, both the removeDupTumor and matchingExpAndMetSamples arguments must be
set to TRUE. Defaults to TRUE.}

\item{matchingExpAndMetSamples}{Select the type of expression and
methylation sample data matching to perform. If set to TRUE, only samples
with at least one methylation and expression sample annotated to their
patient, will be kept. If set to FALSE, all samples will be kept, including
those without matching expression and methylation data. \strong{Note:} To properly
create a dataset for use with TENET, both the removeDupTumor and
matchingExpAndMetSamples arguments must be set to TRUE. Defaults to TRUE.}

\item{clinicalSurvivalData}{Select how clinical data should be prepared
from the TCGA data, with respect to patient vital status and survival time.
Valid options include "bcrBiotabPatient" to use survival data contained
only in the 'patient' data in the BCR Biotab files downloaded using
TCGAbiolinks, or "combined", which uses clinical information from the
'patient' and 'follow_up' datasets in the BCR Biotab files, as well as data
from the BCR XML files. Data from the same patient in each of the datasets
are combined, and the data with the most recent (highest patient survival
time) entry for each patient are kept. Additionally, for both options, the
'days_to_last_followup' and 'days_to_death' variables are collapsed into a
single time variable, which is combined with the other patient clinical data
in the 'patient' BCR Biotab data. See
\url{https://bioconductor.org/packages/devel/bioc/vignettes/TCGAbiolinks/inst/doc/clinical.html}
for more information on how TCGAbiolinks prepares different clinical
datasets. Defaults to "combined".}

\item{outputFile}{Specify the path to an .rda file in which to save the
MultiAssayExperiment object with downloaded datasets. If set to NA or
undefined, this results in the function only returning the
MultiAssayExperiment object and not saving it. Defaults to NA.}
}
\value{
Returns and/or saves to an .rda file a MultiAssayExperiment
object with expression and methylation data included SummarizedExperiment
objects within the MultiAssayExperiment object, as well as clinical data
included in the colData of the MultiAssayExperiment object.
}
\description{
This function downloads and compiles TCGA gene expression and DNA
methylation datasets, as well as clinical data primarily intended for use
with the TENET package. This simplifies the TCGAbiolinks download functions,
identifies samples with matching gene expression and DNA methylation data,
and can also remove duplicate tumor samples taken from the same patient
donor. Data are compiled into a MultiAssayExperiment object, which is
returned and optionally saved in an .rda file at the path specified by the
\code{outputFile} argument.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
## Download a TCGA LUAD dataset with log2-normalized
## FPKM-UQ expression values from tumor and adjacent normal tissue samples
## with matching expression and methylation data and keeping only one tumor
## sample from each patient. Additionally, survival data will be combined
## from three clinical datasets downloaded by TCGAbiolinks. Raw data files
## will be saved to the working directory, and the processed dataset will
## be returned as a variable.
TCGADataset <- TCGADownloader(
    rawDataDownloadDirectory = ".",
    TCGAStudyAbbreviation = "LUAD",
    RNASeqWorkflow = "STAR - FPKM-UQ"
)

## Another example, which downloads a TCGA BRCA dataset with FPKM expression
## values with no normalization and no duplicate samples removed. Survival
## data are derived from just the patient BCR Biotab file downloaded by
## TCGAbiolinks. Both raw data files and an .rda file containing the data
## as a MultiAssayExperiment object will be saved to the working directory.
## Note: This functionality is useful for downloading samples from
## TCGA but will *not* work for a TENET assay due to the lack of sample
## matching and duplicate tumor sample removal.
TCGADownloader(
    rawDataDownloadDirectory = ".",
    TCGAStudyAbbreviation = "BRCA",
    RNASeqWorkflow = "STAR - FPKM",
    RNASeqLog2Normalization = FALSE,
    removeDupTumor = FALSE,
    matchingExpAndMetSamples = FALSE,
    clinicalSurvivalData = "bcrBiotabPatient",
    outputFile = "BRCAMultiAssayExperimentObject.rda"
)
\dontshow{\}) # examplesIf}
}
