% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TENET-package.R
\docType{package}
\name{TENET}
\alias{TENET-package}
\alias{TENET}
\title{TENET (Tracing regulatory Element Networks using Epigenetic Traits)}
\description{
TENET identifies key transcription factors and regulatory
elements linked to a specific cell type by finding significantly correlated
differences in gene expression and regulatory element methylation between
case and control input datasets, and identifying the top genes by number of
significant RE DNA methylation site links. It also includes many additional
tools to aid in the visualization and analysis of the results, including
plots displaying and comparing methylation and expression data and RE DNA
methylation site link counts, survival analysis, TF motif searching in the
vicinity of linked RE DNA methylation sites, custom TAD and peak overlap
analysis, and UCSC Genome Browser track file generation. A utility function
is also provided to download methylation, expression, and patient survival
data from The Cancer Genome Atlas (TCGA) for use in TENET or other analyses.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rhielab/TENET}
  \item Report bugs at \url{https://github.com/rhielab/TENET/issues}
}

}
\author{
\strong{Maintainer}: Rhie Lab at the University of Southern California \email{rhielab@gmail.com}

Authors:
\itemize{
  \item Daniel Mullen (\href{https://orcid.org/0000-0002-7639-0549}{ORCID})
  \item Zexun Wu (\href{https://orcid.org/0000-0003-2566-1326}{ORCID})
  \item Ethan Nelson-Moore (\href{https://orcid.org/0009-0001-6903-9232}{ORCID})
  \item Suhn Rhie (\href{https://orcid.org/0000-0002-5522-5296}{ORCID})
}

}
\keyword{internal}
